app/requirements.txt
urllib3<=1.25.3
Django<=2.2.3
django-filter<=2.1.0
graphene-django
django-cors-headers<=3.0.2
django-choices<=1.7.0
django-jwt-auth<=0.0.2
django-mysql<=3.2.0
ipaddress<=1.0.22
django-redis<=4.10.0
Pillow<6.0.0
mysqlclient<=1.4.2
six<=1.12.0
requests<=2.22.0
regex<=2019.06.08
strsim==0.0.3
analytics-python<=1.2.9
gunicorn<=19.9.0
eventlet<=0.25.1
confusable_homoglyphs<=3.2.0
cloudinary
